poc_files_in_folders/pyproject.toml
[tool.poetry]
name = "poc_files_in_folders"
version = "0.1.0"
description = ""
authors = ["Fernando Sedrez <fmsedrez@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
functions-framework = "^3.2.0"
google-cloud-logging = "^3.2.5"
google-cloud-storage = "^2.5.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
